+++ cmp/ag-grid/AgGridModel.ts
@@ -489,12 +489,17 @@ export class AgGridModel extends HoistModel {
     setSelectedRowNodeIds(ids: string[]) {
         this.throwIfNotReady();
 
-        const {agApi} = this;
-        agApi.deselectAll();
+        const {agApi} = this,
+            idSet = new Set(ids),
+            toDeselect = agApi.getSelectedNodes().filter(it => !idSet.has(it.id)),
+            toSelect: IRowNode[] = [];
         ids.forEach(id => {
             const node = agApi.getRowNode(id);
-            if (node) node.setSelected(true);
+            if (node && !node.isSelected()) toSelect.push(node);
         });
+
+        if (!isEmpty(toDeselect)) agApi.setNodesSelected({nodes: toDeselect, newValue: false});
+        if (!isEmpty(toSelect)) agApi.setNodesSelected({nodes: toSelect, newValue: true});
     }
 
     /**